graceful shutdown reason typed
