+++ huginn-proxy-lib/src/lib.rs
@@ -26,4 +26,4 @@ pub use proxy::reload::{
 pub use proxy::server::{SynProbe, WatchOptions};
 pub use proxy::shutdown::{shutdown_channel, ShutdownPhase, ShutdownSender, ShutdownWatch};
 pub use proxy::{forwarding, run};
-pub use telemetry::{Metrics, Readiness};
+pub use telemetry::{Metrics, NotReadyReason, Readiness};
