+++ huginn-proxy-lib/src/telemetry/health.rs
@@ -2,6 +2,7 @@ use hyper::Response;
 use hyper::StatusCode;
 use tracing::warn;
 
+use crate::telemetry::readiness::NotReadyReason;
 use crate::telemetry::status::{Status, StatusBody};
 use crate::utils::http::{json_response, RespBody};
 
@@ -19,14 +20,14 @@ pub fn live_check_response() -> Response<RespBody> {
 /// accepting connections.
 /// `not_ready_reason` is `None` when ready; otherwise the JSON `reason`
 /// (`proxy_starting` or `proxy_draining`).
-pub fn ready_check_response(not_ready_reason: Option<&'static str>) -> Response<RespBody> {
+pub fn ready_check_response(not_ready_reason: Option<NotReadyReason>) -> Response<RespBody> {
     let Some(reason) = not_ready_reason else {
         return json_response(StatusCode::OK, StatusBody::new(Status::Ready));
     };
 
-    warn!(reason, "Readiness check failed");
+    warn!(reason = reason.as_str(), "Readiness check failed");
     json_response(
         StatusCode::SERVICE_UNAVAILABLE,
-        StatusBody::with_reason(Status::NotReady, reason),
+        StatusBody::with_reason(Status::NotReady, reason.as_str()),
     )
 }
