+++ huginn-proxy-lib/src/telemetry/mod.rs
@@ -10,6 +10,6 @@ pub mod tracing;
 pub use health::{health_check_response, live_check_response, ready_check_response};
 pub use metrics::{init_metrics, values, Metrics};
 pub use metrics_handler::handle_metrics;
-pub use readiness::Readiness;
+pub use readiness::{NotReadyReason, Readiness};
 pub use server::start_observability_server;
 pub use tracing::{init_tracing_with_otel, init_validation_tracing, shutdown_tracing};
